Started the implementation of the Win32 formatting utility; refactoring: made card size int
git-svn-id: https://svn.code.sf.net/p/szte-wsn/code@2912 a4e16f3a-606c-4714-833f-ed604d7aca08

+++ private/ali/RawToCType/Console.hpp
@@ -34,6 +34,7 @@
 #ifndef CONSOLE_HPP_
 #define CONSOLE_HPP_
 
+#include <stdint.h>
 #include "TypeDefs.hpp"
 
 namespace sdc {
@@ -42,9 +43,9 @@ class Console {
 
 public:
 
-	static void start(double card_size, int mote_id, int first_block, int reboot);
+	static void start(int64_t card_size, int mote_id, int first_block, int reboot);
 
-	static void finished(double card_size, int last_block);
+	static void finished(int64_t card_size, int last_block);
 
 	static void record_start(int reboot_id, int first_offset);
